qcacld-3.0: Update he rx/tx MCS variable to hold uint32_t
Update he rx/tx MCS variable to hold uint32_t since WMA layer
expects it to be uint32_t.

CRs-Fixed: 1073481

+++ core/mac/inc/sir_api.h
@@ -368,8 +368,8 @@ typedef struct sSirSupportedRates {
 	uint16_t vhtTxMCSMap;
 	uint16_t vhtTxHighestDataRate;
 #ifdef WLAN_FEATURE_11AX
-	uint16_t he_rx_mcs;
-	uint16_t he_tx_mcs;
+	uint32_t he_rx_mcs;
+	uint32_t he_tx_mcs;
 #endif
 } tSirSupportedRates, *tpSirSupportedRates;
